# implement dpnp.mean