--- a/dpnp/dpnp_array.py
+++ b/dpnp/dpnp_array.py
@@ -915,10 +915,14 @@ class dpnp_array:
 
         return dpnp.max(self, axis, out, keepdims, initial, where)
 
-    def mean(self, axis=None, **kwargs):
-        """Returns the average of the array elements."""
+    def mean(self, axis=None, dtype=None, keepdims=False, out=None, where=True):
+        """
+        Returns the average of the array elements.
 
-        return dpnp.mean(self, axis=axis, **kwargs)
+        Refer to :obj:`dpnp.mean` for full documentation.
+        """
+
+        return dpnp.mean(self, axis, dtype, keepdims, out, where=where)
 
     def min(
         self,