--- a/dpnp/dpnp_array.py
+++ b/dpnp/dpnp_array.py
@@ -915,14 +915,16 @@ class dpnp_array:
 
         return dpnp.max(self, axis, out, keepdims, initial, where)
 
-    def mean(self, axis=None, dtype=None, keepdims=False, out=None, where=True):
+    def mean(
+        self, axis=None, dtype=None, out=None, keepdims=False, *, where=True
+    ):
         """
         Returns the average of the array elements.
 
         Refer to :obj:`dpnp.mean` for full documentation.
         """
 
-        return dpnp.mean(self, axis, dtype, keepdims, out, where=where)
+        return dpnp.mean(self, axis, dtype, out, keepdims, where=where)
 
     def min(
         self,