--- a/tests/test_statistics.py
+++ b/tests/test_statistics.py
@@ -8,7 +8,7 @@ from numpy.testing import (
 
 import dpnp
 
-from .helper import get_all_dtypes
+from .helper import assert_dtype_allclose, get_all_dtypes
 
 
 @pytest.mark.parametrize(
@@ -86,6 +86,73 @@ def test_max_min_NotImplemented(func):
         getattr(dpnp, func)(ia, where=False)
     with pytest.raises(NotImplementedError):
         getattr(dpnp, func)(ia, initial=6)
+
+
+class TestMean:
+    @pytest.mark.parametrize("dtype", get_all_dtypes())
+    def test_mean_axis_tuple(self, dtype):
+        dp_array = dpnp.array([[0, 1, 2], [3, 4, 0]], dtype=dtype)
+        np_array = dpnp.asnumpy(dp_array)
+
+        result = dpnp.mean(dp_array, axis=(0, 1))
+        expected = numpy.mean(np_array, axis=(0, 1))
+        assert_allclose(expected, result)
+
+    @pytest.mark.parametrize("dtype", get_all_dtypes())
+    @pytest.mark.parametrize("axis", [0, 1, (0, 1)])
+    def test_mean_out(self, dtype, axis):
+        dp_array = dpnp.array([[0, 1, 2], [3, 4, 0]], dtype=dtype)
+        np_array = dpnp.asnumpy(dp_array)
+
+        expected = numpy.mean(np_array, axis=axis)
+        result = dpnp.empty_like(dpnp.asarray(expected))
+        dpnp.mean(dp_array, axis=axis, out=result)
+        assert_dtype_allclose(result, expected)
+
+    @pytest.mark.parametrize("dtype", get_all_dtypes())
+    def test_mean_dtype(self, dtype):
+        dp_array = dpnp.array([[0, 1, 2], [3, 4, 0]], dtype="i4")
+        np_array = dpnp.asnumpy(dp_array)
+
+        expected = numpy.mean(np_array, dtype=dtype)
+        result = dpnp.mean(dp_array, dtype=dtype)
+        assert_allclose(expected, result)
+
+    @pytest.mark.usefixtures("suppress_invalid_numpy_warnings")
+    @pytest.mark.parametrize("axis", [0, 1, (0, 1)])
+    @pytest.mark.parametrize("shape", [(2, 3), (2, 0), (0, 3)])
+    def test_mean_empty(self, axis, shape):
+        dp_array = dpnp.empty(shape, dtype=dpnp.int64)
+        np_array = dpnp.asnumpy(dp_array)
+
+        result = dpnp.mean(dp_array, axis=axis)
+        expected = numpy.mean(np_array, axis=axis)
+        assert_allclose(expected, result)
+
+    def test_mean_strided(self):
+        dp_array = dpnp.array([-2, -1, 0, 1, 0, 2], dtype="f4")
+        np_array = dpnp.asnumpy(dp_array)
+
+        result = dpnp.mean(dp_array[::-1])
+        expected = numpy.mean(np_array[::-1])
+        assert_allclose(expected, result)
+
+        result = dpnp.mean(dp_array[::2])
+        expected = numpy.mean(np_array[::2])
+        assert_allclose(expected, result)
+
+    def test_mean_scalar(self):
+        dp_array = dpnp.array(5)
+        np_array = dpnp.asnumpy(dp_array)
+
+        result = dp_array.mean()
+        expected = np_array.mean()
+        assert_allclose(expected, result)
+
+    def test_mean_NotImplemented(func):
+        ia = dpnp.arange(5)
+        with pytest.raises(NotImplementedError):
+            dpnp.mean(ia, where=False)
 
 
 @pytest.mark.usefixtures("allow_fall_back_on_numpy")