@vtavana 2023-11-29
Performance on PVC (Intel(R) Data Center GPU Max 1550):
Timing is obtained using `%timeit -r 7 -n 100` for three different cases:

1. `a = dpnp.arange(1024*32).reshape(1024,32)`

| Function | Old implementation | New Implementation |
| :---         |     :---:      |    :---:    |
| `mean(a)`   | 1.98 ms ± 128  µs    | 1.28 ms ± 44.4 µs   |
| `mean(a, axis=0)`   | 2.39 ms ± 26.6 µs       | 1.52 ms ± 907 µs      |
| `mean(a, axis=1)`   | 1.97 ms ± 269 µs      | 1.37 ms ± 494 µs      |

2.  `a = dpnp.arange(1024*32).reshape(32,1024)`

| Function | Old implementation | New Implementation |
| :---         |     :---:      |    :---:    |
| `mean(a)`   | 1.89 ms ± 150  µs    | 1.3 ms ± 33 µs   |
| `mean(a, axis=0)`   | 2.41 ms ± 48.3 µs       | 1.17 ms ± 34.1 µs      |
| `mean(a, axis=1)`   | 2.33 ms ± 135 µs      | 1.18 ms ± 26.9 µs      |

3.  `a = dpnp.arange(300*300).reshape(300,300)`

| Function | Old implementation | New Implementation |
| :---         |     :---:      |    :---:    |
| `mean(a)`   | 2.21 ms ± 78.3  µs    | 1.29 ms ± 58.5 µs   |
| `mean(a, axis=0)`   | 2.4 ms ± 28.9 µs       | 1.15 ms ± 29.8 µs      |
| `mean(a, axis=1)`   | 2.39 ms ± 30 µs      | 1.16 ms ± 34 µs      |