@vtavana 2023-11-29
Performance on Iris Xe (Intel(R) Graphics [0x46a8]):
Timing is obtained using `%timeit -r 7 -n 100` for three different cases:

1. `a = dpnp.arange(1024*32).reshape(1024,32)`

| Function | Old implementation | New Implementation |
| :---         |     :---:      |    :---:    |
| `mean(a)`   | 2.13 ms ± 265  µs    | 621 µs ± 93 µs   |
| `mean(a, axis=0)`   | 1.96 ms ± 127 µs       | 673 µs ± 102 µs      |
| `mean(a, axis=1)`   | 1.97 ms ± 122 µs      | 674 µs ± 71.4 µs      |

2.  `a = dpnp.arange(1024*32).reshape(32,1024)`

| Function | Old implementation | New Implementation |
| :---         |     :---:      |    :---:    |
| `mean(a)`   | 1.9 ms ± 101  µs    | 574 µs ± 87 µs   |
| `mean(a, axis=0)`   | 1.99 ms ± 55.3 µs       | 737 µs ± 98.2 µs      |
| `mean(a, axis=1)`   | 1.97 ms ± 105 µs      | 656 µs ± 85.1 µs      |

3.  `a = dpnp.arange(300*300).reshape(300,300)`

| Function | Old implementation | New Implementation |
| :---         |     :---:      |    :---:    |
| `mean(a)`   | 1.92 ms ± 147  µs    | 622 µs ± 74.5 µs   |
| `mean(a, axis=0)`   | 1.98 ms ± 113 µs       | 654 µs ± 103 µs      |
| `mean(a, axis=1)`   | 1.96 ms ± 108 µs      | 698 µs ± 122 µs      |